chatbot/requirements.txt
langchain
streamlit==1.18.1
faiss-cpu
streamlit-extras
openai
pinecone-client
tiktoken